# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "Dokumentasi API e-commerce MK3",
        "contact": {
            "email": "kamu@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/register": {
            "post": {
                "tags": [
                    "Auth - Admin"
                ],
                "summary": "Register admin baru",
                "operationId": "d9d33b06557dd8387b91408f67990a0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin Baru"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil register"
                    },
                    "500": {
                        "description": "Gagal register"
                    }
                }
            }
        },
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Auth - Admin"
                ],
                "summary": "Login admin dan dapatkan token",
                "operationId": "36a65807d999036719cfae8de188aebe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/admin/get": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Ambil semua admin",
                "operationId": "4275198b3ed869955314c777bc48f091",
                "responses": {
                    "200": {
                        "description": "Daftar admin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/logout": {
            "post": {
                "tags": [
                    "Auth - Admin"
                ],
                "summary": "Logout dan hapus semua token",
                "operationId": "6af39a0428937177fbdc1567f67f7ea0",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/delete": {
            "delete": {
                "tags": [
                    "Auth - Admin"
                ],
                "summary": "Hapus akun admin",
                "operationId": "2fc3cbd509d4c56e58a6b2dcb0eba094",
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus akun"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/update-profile": {
            "put": {
                "tags": [
                    "Auth - Admin"
                ],
                "summary": "Update profile admin",
                "operationId": "72063b107b1f48a9391863798adca02e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nama Baru"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "emailbaru@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/register": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Register a new customer",
                "operationId": "784859d9cb00e95f0932a00529e02263",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Mawar No. 1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer registered successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Registration failed"
                    }
                }
            }
        },
        "/api/customer/login": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Login customer and get token",
                "operationId": "771b3016c553bffac97f68c8a9babad9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Login failed"
                    },
                    "500": {
                        "description": "Login failed"
                    }
                }
            }
        },
        "/api/customer/profile": {
            "get": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Get authenticated customer profile",
                "operationId": "75a34bd8cbfca11fbd7b44b528eed47d",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    },
                    "500": {
                        "description": "Failed to get profile"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/update-profile": {
            "put": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Update authenticated customer profile",
                "operationId": "db3da107ad39a984920e9ca72e8dee43",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Anggrek No. 2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Profile update failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/logout": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Logout authenticated customer",
                "operationId": "b2ea3e5654430383466e91b256e2cc31",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "500": {
                        "description": "Logout failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer Token via Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}